move internal stuff out of header
